requirements.txt
#devops
black ==22.3.0 # Formatter
click == 8.1.3
pytest == 7.4.0 #For Testing
pytest-cov == 4.0.0 #For Testing
ruff==0.0.284 #For linting
pandas == 2.2.2
numpy == 1.26.4

#web
fastapi == 0.85.0
uvicorn == 0.18.3

#Databricks and Python-dotenv
databricks-sql-connector == 2.0.2
python-dotenv == 1.0.1